Update the structure of lat and longs

+++ database/migrations/2017_03_15_194901_create_offer__locations_table.php
@@ -19,8 +19,8 @@ class CreateOfferLocationsTable extends Migration
             $table->string('name')->nullable();
             $table->string('address')->nullable();
             $table->integer('offer_id');
-            $table->decimal('lat')->nullable();
-            $table->decimal('lng')->nullable();
+            $table->double('lat',11,8)->nullable();
+            $table->double('lng',11,8)->nullable();
             $table->string('type')->nullable();
             $table->boolean('countrywide')->nullable();
         });
